\name{tritrypdb-package}
\alias{tritrypdb-package}
\alias{tritrypdb}
\docType{package}
\title{
R functions for TriTrypDB.
}
\description{
A small collection of R functions for working with TriTrypDB data files.
}
\details{
\tabular{ll}{
Package: \tab tritrypdb\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-06-24\cr
License: \tab BSD_3_clause\cr
}
}
\author{
Keith Hughitt<\url{http://github.com/khughitt}>
}
\references{
  TriTrypDB: \url{http://tritrypdb.org/tritrypdb/}
}
